+++ state/execution.go
@@ -7,23 +7,25 @@ import (
 )
 
 func (st *state) executeBlock(block block.Block) ([]tx.CommittedTx, error) {
+	st.txPoolSandbox.Clear()
+	st.executionSandbox.Clear()
+	st.execution.ResetFee()
+
 	ids := block.TxIDs().IDs()
 	twrs := make([]tx.CommittedTx, len(ids))
-
+	var subsidyTrx *tx.Tx
 	for i := 0; i < len(ids); i++ {
 		trx := st.txPool.PendingTx(ids[i])
 		if trx == nil {
 			return nil, errors.Errorf(errors.ErrInvalidBlock, "Transaction not found")
 		}
-		if err := trx.SanityCheck(); err != nil {
-			return nil, err
-		}
 		// Only first transaction should be subsidy transaction
 		isSubsidyTx := (i == 0)
 		if isSubsidyTx {
 			if !trx.IsSubsidyTx() {
-				return nil, errors.Errorf(errors.ErrInvalidTx, "Not a subsidy transaction")
+				return nil, errors.Errorf(errors.ErrInvalidTx, "First transaction should be a subsidy transaction")
 			}
+			subsidyTrx = trx
 		} else {
 			if trx.IsSubsidyTx() {
 				return nil, errors.Errorf(errors.ErrInvalidTx, "Duplicated subsidy transaction")
@@ -37,6 +39,11 @@ func (st *state) executeBlock(block block.Block) ([]tx.CommittedTx, error) {
 		receipt := trx.GenerateReceipt(tx.Ok, block.Hash())
 		twrs[i].Tx = trx
 		twrs[i].Receipt = receipt
+	}
+
+	subsidyAmt := calcBlockSubsidy(st.lastBlockHeight+1, st.params.SubsidyReductionInterval) + st.execution.AccumulatedFee()
+	if subsidyTrx.Payload().Value() != subsidyAmt {
+		return nil, errors.Errorf(errors.ErrInvalidTx, "Invalid subsidy amount. Expected %v, got %v", subsidyAmt, subsidyTrx.Payload().Value())
 	}
 
 	return twrs, nil