+++ state/mock.go
@@ -15,9 +15,10 @@ import (
 var _ State = &MockState{}
 
 type MockState struct {
-	LastBlockCommit *block.Commit
-	GenHash         crypto.Hash
-	Store           *store.MockStore
+	LastBlockCommit  *block.Commit
+	GenHash          crypto.Hash
+	Store            *store.MockStore
+	InvalidBlockHash crypto.Hash
 }
 
 func NewMockStore() *MockState {
@@ -51,13 +52,17 @@ func (m *MockState) LastCommit() *block.Commit {
 func (m *MockState) BlockTime() time.Duration {
 	return time.Second
 }
-func (m *MockState) UpdateLastCommit(blockHash crypto.Hash, commit block.Commit) {
-	m.LastBlockCommit = &commit
+func (m *MockState) UpdateLastCommit(commit *block.Commit) error {
+	m.LastBlockCommit = commit
+	return nil
 }
 func (m *MockState) Fingerprint() string {
 	return ""
 }
 func (m *MockState) ApplyBlock(height int, b block.Block, c block.Commit) error {
+	if b.Hash().EqualsTo(m.InvalidBlockHash) {
+		return fmt.Errorf("Invalid block")
+	}
 	if height == m.LastBlockHeight()+1 {
 		m.Store.Blocks[height] = &b
 		m.LastBlockCommit = &c