+++ state/state_test.go
@@ -4,82 +4,107 @@ import (
 	"testing"
 	"time"
 
-	"github.com/zarbchain/zarb-go/tx"
-
-	"github.com/zarbchain/zarb-go/tx/payload"
-
 	"github.com/stretchr/testify/assert"
-
 	"github.com/stretchr/testify/require"
 	"github.com/zarbchain/zarb-go/account"
 	"github.com/zarbchain/zarb-go/block"
 	"github.com/zarbchain/zarb-go/crypto"
 	"github.com/zarbchain/zarb-go/genesis"
 	"github.com/zarbchain/zarb-go/logger"
+	"github.com/zarbchain/zarb-go/tx/payload"
 	"github.com/zarbchain/zarb-go/txpool"
 	"github.com/zarbchain/zarb-go/util"
 	"github.com/zarbchain/zarb-go/validator"
 	"github.com/zarbchain/zarb-go/vote"
 )
 
-var tTxPool *txpool.MockTxPool
-var tGenDoc *genesis.Genesis
-var tValSigner crypto.Signer
+var tValSigner1 crypto.Signer
+var tValSigner2 crypto.Signer
+var tValSigner3 crypto.Signer
+var tValSigner4 crypto.Signer
+var tGenTime time.Time
+var tCommonTxPool *txpool.MockTxPool
 
-func setup(t *testing.T) {
-	tTxPool = txpool.NewMockTxPool()
+func init() {
+	_, _, priv1 := crypto.GenerateTestKeyPair()
+	_, _, priv2 := crypto.GenerateTestKeyPair()
+	_, _, priv3 := crypto.GenerateTestKeyPair()
+	_, _, priv4 := crypto.GenerateTestKeyPair()
 
-	_, pb, priv := crypto.GenerateTestKeyPair()
-	acc := account.NewAccount(crypto.TreasuryAddress, 0)
-	acc.AddToBalance(21000000000000)
-	val := validator.NewValidator(pb, 0, 0)
-	val.AddToStake(util.RandInt64(10000000))
-	tGenDoc = genesis.MakeGenesis("test", time.Now(), []*account.Account{acc}, []*validator.Validator{val}, 1)
-	tValSigner = crypto.NewSigner(priv)
+	tValSigner1 = crypto.NewSigner(priv1)
+	tValSigner2 = crypto.NewSigner(priv2)
+	tValSigner3 = crypto.NewSigner(priv3)
+	tValSigner4 = crypto.NewSigner(priv4)
+
+	tGenTime = util.Now()
 
 	loggerConfig := logger.TestConfig()
 	logger.InitLogger(loggerConfig)
+
+	tCommonTxPool = txpool.NewMockTxPool()
 }
 
-func mockState(t *testing.T, signer *crypto.Signer) (*state, crypto.Address) {
-	if signer == nil {
-		_, _, priv := crypto.GenerateTestKeyPair()
-		s := crypto.NewSigner(priv)
-		signer = &s
-	}
+func setupStatewithFourValidators(t *testing.T, signer crypto.Signer) *state {
 
-	stateConfig := TestConfig()
-	st, err := LoadOrNewState(stateConfig, tGenDoc, *signer, tTxPool)
+	acc := account.NewAccount(crypto.TreasuryAddress, 0)
+	// 2,100,000,000,000,000
+	acc.AddToBalance(21 * 1e14)
+	val1 := validator.NewValidator(tValSigner1.PublicKey(), 0, 0)
+	val2 := validator.NewValidator(tValSigner2.PublicKey(), 1, 0)
+	val3 := validator.NewValidator(tValSigner3.PublicKey(), 2, 0)
+	val4 := validator.NewValidator(tValSigner4.PublicKey(), 3, 0)
+	gnDoc := genesis.MakeGenesis("test", tGenTime, []*account.Account{acc}, []*validator.Validator{val1, val2, val3, val4}, 1)
+
+	st, err := LoadOrNewState(TestConfig(), gnDoc, signer, tCommonTxPool)
 	require.NoError(t, err)
 	s, _ := st.(*state)
 
-	return s, signer.Address()
+	return s
+}
+
+func setupStatewithOneValidator(t *testing.T) *state {
+	acc := account.NewAccount(crypto.TreasuryAddress, 0)
+	acc.AddToBalance(21 * 1e14)
+	val := validator.NewValidator(tValSigner1.PublicKey(), 0, 0)
+	genDoc := genesis.MakeGenesis("test", tGenTime, []*account.Account{acc}, []*validator.Validator{val}, 1)
+
+	st, err := LoadOrNewState(TestConfig(), genDoc, tValSigner1, txpool.NewMockTxPool())
+	require.NoError(t, err)
+	s, _ := st.(*state)
+
+	return s
 }
 
 func TestProposeBlockAndValidation(t *testing.T) {
-	setup(t)
+	st := setupStatewithOneValidator(t)
 
-	st, _ := mockState(t, &tValSigner)
 	block := st.ProposeBlock()
 	err := st.ValidateBlock(block)
 	require.NoError(t, err)
 }
 
-func proposeAndSignBlock(t *testing.T, st *state) (block.Block, block.Commit) {
-	addr := tValSigner.Address()
+func proposeAndSignBlock(t *testing.T, st *state, signer crypto.Signer) (block.Block, block.Commit) {
 	b := st.ProposeBlock()
-	v := vote.NewPrecommit(1, 0, b.Hash(), addr)
-	sig := tValSigner.Sign(v.SignBytes())
-	c := block.NewCommit(0, []block.Committer{{Status: 1, Address: addr}}, *sig)
+	c := makeCommitAndSign(t, b.Hash(), signer)
 
-	return b, *c
+	return b, c
+}
+
+func makeCommitAndSign(t *testing.T, blockHash crypto.Hash, signers ...crypto.Signer) block.Commit {
+	committers := make([]block.Committer, len(signers))
+	sigs := make([]*crypto.Signature, len(signers))
+	for i, s := range signers {
+		v := vote.NewPrecommit(-1, 0, blockHash, s.Address())
+
+		committers[i] = block.Committer{Status: 1, Address: s.Address()}
+		sigs[i] = s.Sign(v.SignBytes())
+	}
+	return *block.NewCommit(0, committers, crypto.Aggregate(sigs))
 }
 
 func TestLoadState(t *testing.T) {
-	setup(t)
-
-	st1, _ := mockState(t, &tValSigner)
-	st2, _ := mockState(t, &tValSigner)
+	st1 := setupStatewithOneValidator(t)
+	st2 := setupStatewithOneValidator(t)
 
 	// Add this dummy acc and val for testing purpose
 	dummyAcc, _ := account.GenerateTestAccount(1)
@@ -93,19 +118,22 @@ func TestLoadState(t *testing.T) {
 
 	i := 0
 	for ; i < 10; i++ {
-		b, c := proposeAndSignBlock(t, st1)
+		b1, c1 := proposeAndSignBlock(t, st1, tValSigner1)
+		b2, c2 := proposeAndSignBlock(t, st2, tValSigner1)
 
-		assert.NoError(t, st1.ApplyBlock(i+1, b, c))
-		assert.NoError(t, st2.ApplyBlock(i+1, b, c))
+		assert.NoError(t, st1.ApplyBlock(i+1, b1, c1))
+		assert.NoError(t, st2.ApplyBlock(i+1, b2, c2))
 	}
+
+	assert.Equal(t, st1.stateHash(), st2.stateHash())
 
 	assert.NoError(t, st2.Close())
 
 	// Load last state info
-	st3, err := LoadOrNewState(st2.config, tGenDoc, tValSigner, tTxPool)
+	st3, err := LoadOrNewState(st2.config, st2.genDoc, tValSigner1, txpool.NewMockTxPool())
 	require.NoError(t, err)
 
-	b, c := proposeAndSignBlock(t, st1)
+	b, c := proposeAndSignBlock(t, st1, tValSigner1)
 	assert.Equal(t, b.Hash(), st3.ProposeBlock().Hash())
 	require.NoError(t, st1.ApplyBlock(i+1, b, c))
 	require.NoError(t, st3.ApplyBlock(i+1, b, c))
@@ -117,24 +145,21 @@ func TestLoadState(t *testing.T) {
 }
 
 func TestBlockSubsidy(t *testing.T) {
-	setup(t)
-
-	interval := 210000
-	assert.Equal(t, int64(5*1e8), calcBlockSubsidy(1, 210000))
-	assert.Equal(t, int64(5*1e8), calcBlockSubsidy((1*interval)-1, 210000))
-	assert.Equal(t, int64(2.5*1e8), calcBlockSubsidy((1*interval), 210000))
-	assert.Equal(t, int64(2.5*1e8), calcBlockSubsidy((2*interval)-1, 210000))
-	assert.Equal(t, int64(1.25*1e8), calcBlockSubsidy((2*interval), 210000))
+	interval := 2100000
+	assert.Equal(t, int64(5*1e8), calcBlockSubsidy(1, interval))
+	assert.Equal(t, int64(5*1e8), calcBlockSubsidy((1*interval)-1, interval))
+	assert.Equal(t, int64(2.5*1e8), calcBlockSubsidy((1*interval), interval))
+	assert.Equal(t, int64(2.5*1e8), calcBlockSubsidy((2*interval)-1, interval))
+	assert.Equal(t, int64(1.25*1e8), calcBlockSubsidy((2*interval), interval))
 }
 
 func TestBlockSubsidyTx(t *testing.T) {
-	setup(t)
-	st, _ := mockState(t, &tValSigner)
+	st := setupStatewithOneValidator(t)
 
 	trx := st.createSubsidyTx(7)
 	assert.True(t, trx.IsSubsidyTx())
 	assert.Equal(t, trx.Payload().Value(), calcBlockSubsidy(1, st.params.SubsidyReductionInterval)+7)
-	assert.Equal(t, trx.Payload().(*payload.SendPayload).Receiver, tValSigner.Address())
+	assert.Equal(t, trx.Payload().(*payload.SendPayload).Receiver, tValSigner1.Address())
 
 	addr, _, _ := crypto.GenerateTestKeyPair()
 	st.config.MintbaseAddress = &addr
@@ -143,44 +168,11 @@ func TestBlockSubsidyTx(t *testing.T) {
 }
 
 func TestApplyBlocks(t *testing.T) {
-	setup(t)
+	st := setupStatewithOneValidator(t)
 
-	st, _ := mockState(t, &tValSigner)
-	b1, c1 := proposeAndSignBlock(t, st)
+	b1, c1 := proposeAndSignBlock(t, st, tValSigner1)
 	invBlock, _ := block.GenerateTestBlock(nil)
 	assert.Error(t, st.ApplyBlock(1, *invBlock, c1))
 	assert.Error(t, st.ApplyBlock(2, b1, c1))
-}
-
-func TestProposeBlock(t *testing.T) {
-	setup(t)
-
-	st1, _ := mockState(t, &tValSigner)
-	st2, _ := mockState(t, nil)
-
-	b1, c1 := proposeAndSignBlock(t, st1)
-	assert.NoError(t, st1.ApplyBlock(1, b1, c1))
-
-	invSubsidyTx := st2.createSubsidyTx(100)
-	invSendTx, _ := tx.GenerateTestSendTx()
-	invBondTx, _ := tx.GenerateTestBondTx()
-	invSortitionTx, _ := tx.GenerateTestSortitionTx()
-
-	pub := tValSigner.PublicKey()
-	trx1 := tx.NewSendTx(b1.Hash(), 1, tValSigner.Address(), tValSigner.Address(), 1, 1000, "", &pub, nil)
-	tValSigner.SignMsg(trx1)
-
-	trx2 := tx.NewBondTx(b1.Hash(), 2, tValSigner.Address(), pub, 1, "", &pub, nil)
-	tValSigner.SignMsg(trx2)
-
-	assert.NoError(t, tTxPool.AppendTx(invSendTx))
-	assert.NoError(t, tTxPool.AppendTx(invBondTx))
-	assert.NoError(t, tTxPool.AppendTx(invSortitionTx))
-	assert.NoError(t, tTxPool.AppendTx(invSubsidyTx))
-	assert.NoError(t, tTxPool.AppendTx(trx1))
-	assert.NoError(t, tTxPool.AppendTx(trx2))
-
-	b2 := st1.ProposeBlock()
-	assert.Equal(t, b2.Header().LastBlockHash(), b1.Hash())
-	assert.Equal(t, b2.TxIDs().IDs()[1:], []crypto.Hash{trx1.ID(), trx2.ID()})
+	assert.NoError(t, st.ApplyBlock(1, b1, c1))
 }