+++ vote/vote_set.go
@@ -80,10 +80,6 @@ func (vs *VoteSet) AddVote(vote *Vote) (bool, error) {
 	signer := vote.Signer()
 	blockHash := vote.BlockHash()
 
-	if signer.SanityCheck() != nil {
-		return false, errors.Errorf(errors.ErrInvalidVote, "Empty address")
-	}
-
 	if (vote.data.Height != vs.height) ||
 		(vote.data.Round != vs.round) ||
 		(vote.data.VoteType != vs.voteType) {
@@ -180,14 +176,14 @@ func (vs *VoteSet) ToCommit() *block.Commit {
 	votesMap := vs.votesByBlock[*blockHash].votes
 	vals := vs.valSet.Validators()
 	committers := make([]block.Committer, len(vals))
-	sigs := make([]crypto.Signature, 0)
+	sigs := make([]*crypto.Signature, 0)
 
 	for i, addr := range vals {
 		status := block.CommitNotSigned
 		v := votesMap[addr]
 
 		if v != nil {
-			sigs = append(sigs, *v.Signature())
+			sigs = append(sigs, v.Signature())
 			status = block.CommitSigned
 		}