+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 1.0.0
+- Using blake2b for hashing
+- Using ARC cache for syncing module
+- Reap transaction from txpool before proposing block
+- Executing transaction
+- Updating capnp and http server
+- Making sandbox more isolate
+- Updating state merkle tree
+- Add number to account and validator structure (Used for Merklizing state)
+- Add chain params to genesis
+- Stamp validation check for transaction
+- Refactoring transactions and transaction receipts
+
 ## 0.9.1
 - Adding wizard argument to start command
 - Ensure messages belongs to same network
@@ -9,8 +22,8 @@
 - Refactoring Commit structure
 - Aggregating validators' signatures
 - Validating Commit and committers
-- Merklizing committers
-- Repost UndefHash as a sanity error
+- Calculating root hash of committers
+- Report UndefHash as a sanity error
 - Try to load last state info upon starting the node
 - Saving the last state info when a new block is committed
 - Updating store interface